Split LS and GS. Dont forget update start scripts. DP update required.
git-svn-id: http://svn.l2jserver.com/trunk@4043 d34dfa0f-f91c-0410-86dd-cf06f19dd0ad

+++ L2_GameServer/java/com/l2jserver/gameserver/model/actor/stat/PlayableStat.java
@@ -191,7 +191,7 @@ public class PlayableStat extends CharStat
 			L2ZoneType zone = ZoneManager.getInstance().getZone(getActiveChar(), L2SwampZone.class);
 			int bonus = zone == null ? 0 : ((L2SwampZone) zone).getMoveBonus();
 			double dbonus = bonus / 100.0; //%
-			val += (int) val * dbonus;
+			val += val * dbonus;
 		}
 		
 		return val;
